moved extension classes into Parser file
