+++ Arcsecond/Arcsecond.csproj
@@ -43,17 +43,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BetweenParser.cs" />
-    <Compile Include="ChoiceParser.cs" />
-    <Compile Include="DigitsParser.cs" />
     <Compile Include="Extensions.cs" />
-    <Compile Include="LettersParser.cs" />
-    <Compile Include="ManyParser.cs" />
     <Compile Include="Parser.cs" />
     <Compile Include="ParserState.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SequenceOfParser.cs" />
-    <Compile Include="StringParser.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
