+++ Arcsecond/Parser.cs
@@ -1,9 +1,14 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Text.RegularExpressions;
 
 namespace Arcsecond
 {
     public partial class Parser
     {
+        private static readonly Regex digitsRegex = new Regex("^[0-9]+");
+        private static readonly Regex lettersRegex = new Regex("^[A-Za-z]+");
+
         public readonly Func<ParserState, ParserState> Transform;
 
         public Parser(Func<ParserState, ParserState> transform)
@@ -25,6 +30,7 @@ namespace Arcsecond
             return ParserState.SetResult(state, result.Result, result.Index);
         }
 
+        #region Methods
         public Parser Chain(Func<object, Parser> transform)
         {
             return new Parser(delegate (ParserState state)
@@ -66,5 +72,159 @@ namespace Arcsecond
                 return ParserState.SetError(nextState, transformedError);
             });
         }
+        #endregion
+
+        #region Extension Classes
+        public static Func<Parser, Parser> Between(Parser left, Parser right) => delegate (Parser content)
+        {
+            return SequenceOf(new Parser[] { left, content, right })
+                .Map((results) => ((List<object>)results)[1]);
+        };
+
+        public static Parser Choice(IEnumerable<Parser> parsers) => new Parser(delegate (ParserState state)
+        {
+            if (state.IsError)
+            {
+                return state;
+            }
+
+            foreach (var parser in parsers)
+            {
+                var nextState = parser.Transform(state);
+
+                if (!nextState.IsError)
+                {
+                    return nextState;
+                }
+            }
+
+            return ParserState.SetError(state, $"Unable to match with any parser at index {state.Index}");
+        });
+
+        public static readonly Parser Digits = new Parser(delegate (ParserState state)
+        {
+            if (state.IsError)
+            {
+                return state;
+            }
+
+            var slicedInput = state.Input.Slice(state.Index);
+
+            if (slicedInput.Length == 0)
+            {
+                return ParserState.SetError(state, $"Got unexpected end of input");
+            }
+
+            var match = digitsRegex.Match(slicedInput);
+
+            if (match.Success)
+            {
+                return ParserState.SetResult(state, match.Value, state.Index + match.Value.Length);
+            }
+
+            return ParserState.SetError(state, $"Could not match digits at index {state.Index}");
+        });
+
+        public static readonly Parser Letters = new Parser(delegate (ParserState state)
+        {
+            if (state.IsError)
+            {
+                return state;
+            }
+
+            var slicedInput = state.Input.Slice(state.Index);
+
+            if (slicedInput.Length == 0)
+            {
+                return ParserState.SetError(state, $"Got unexpected end of input");
+            }
+
+            var match = lettersRegex.Match(slicedInput);
+
+            if (match.Success)
+            {
+                return ParserState.SetResult(state, match.Value, state.Index + match.Value.Length);
+            }
+
+            return ParserState.SetError(state, $"Could not match letters at index {state.Index}");
+        });
+
+        public static Parser ManyAtLeast(int minimum, Parser parser) => new Parser(delegate (ParserState state)
+        {
+            if (state.IsError)
+            {
+                return state;
+            }
+
+            var results = new List<object>();
+            var nextState = state;
+
+            while (true)
+            {
+                var testState = parser.Transform(nextState);
+
+                if (!testState.IsError)
+                {
+                    results.Add(testState.Result);
+                    nextState = testState;
+                }
+                else
+                {
+                    break;
+                }
+            }
+
+            if (results.Count < minimum)
+            {
+                return ParserState.SetError(nextState, $"Unable to match any input using parser at index {nextState.Index}");
+            }
+
+            return ParserState.SetResult(nextState, results);
+        });
+
+        public static Parser Many(Parser parser) => ManyAtLeast(0, parser);
+
+        public static Parser SequenceOf(IEnumerable<Parser> parsers) => new Parser(delegate (ParserState state)
+        {
+            if (state.IsError)
+            {
+                return state;
+            }
+
+            var results = new List<object>();
+            var nextState = state;
+
+            foreach (var parser in parsers)
+            {
+                nextState = parser.Transform(nextState);
+
+                results.Add(nextState.Result);
+            }
+
+            return ParserState.SetResult(nextState, results);
+        });
+
+        public static Parser String(string target) => new Parser(delegate (ParserState state)
+        {
+            if (state.IsError)
+            {
+                return state;
+            }
+
+            var slicedInput = state.Input.Slice(state.Index);
+
+            if (slicedInput.Length == 0)
+            {
+                return ParserState.SetError(state, $"Tried to match '{target}', but got unexpected end of input");
+            }
+
+            if (slicedInput.StartsWith(target))
+            {
+                return ParserState.SetResult(state, target, state.Index + target.Length);
+            }
+
+            return ParserState.SetError(state, $"Tried to match '{target}', but got '{state.Input.Slice(state.Index, target.Length)}'");
+        });
+        #endregion
     }
 }
